requirements.txt
PyQt6==6.10.2
PyAudioWPatch==0.2.12.8
cryptography==46.0.6
